Refactor/move test timePickerWithDifferentStep to not use demo view
